Add l3_cache to caps
caps will container a new filed `l3_cache` for each cell.

l3_cache = l3 cache(from /proc/cpuinfo) / cell number.

And a bug fix when get l3_cache for nodeinfo

+++ include/libvirt/libvirt-host.h
@@ -167,7 +167,7 @@ struct _virNodeInfo {
                              processors in case of unusual NUMA topology*/
     unsigned int threads; /* number of threads per core, 1 in case of
                              unusual numa topology */
-    unsigned int l3_cache; /* cache size in kilobytes */
+    unsigned long long l3_cache; /* cache size in kilobytes */
 };
 
 /**
