+++ src/util/virhostcpupriv.h
@@ -35,7 +35,7 @@ int virHostCPUGetInfoPopulateLinux(FILE *cpuinfo,
                                    unsigned int *sockets,
                                    unsigned int *cores,
                                    unsigned int *threads,
-                                   unsigned int *l3_cache);
+                                   unsigned long long *l3_cache);
 
 int virHostCPUGetStatsLinux(FILE *procstat,
                             int cpuNum,
